fix(chat-room): reset the conversation surface when the route id changes
Five findings. One is a data-integrity bug the others are not close to.

SvelteKit REUSES a page component when only a route parameter changes, so a
client-side `/server-owned/A` → `/server-owned/B` updated `data` while the
one-time mirror seed kept A's transcript. The heading and the transport URL
named B while `<Chat>` rendered A — and the next submission would have been
persisted to B underneath A's visible history. Measured, not theorised:
without the fix the second conversation renders "Only the first conversation
has this line".

Latent rather than live today, and the test says so: nothing links one
conversation to another, and browser back between two separately-loaded
documents is a real navigation rather than a routed one. It goes live the
moment any detail-to-detail link exists. The fix is a `{#key data.id}` around
a new `conversation-surface.svelte` holding the mirror, the streaming flag,
the failure banner, and the controller — a key rather than an effect that
resets each field, so the reset is exhaustive by construction instead of
correct only while someone remembers to extend it.

The regression test drives the router's real path — a same-origin anchor
click, which is what SvelteKit intercepts — and asserts a `window` marker
survives it, so a full reload could not make it pass for the wrong reason.
`page.goBack()` was tried first and turns out to be a genuine document
navigation between two separately-loaded entries, which would have proved
nothing.

The signal marker proved disposal STARTED, not that it finished: written
synchronously by the only teardown, it appeared before
`disposeServerOwnedRuntime()` reached its first `await`, so
`void dispose(); process.exit(...)` would still have produced it. The teardown
now crosses a macrotask before writing — a timer rather than
`Promise.resolve()`, because `process.exit` runs after the microtask
checkpoint. That exact regression now fails both cases.

`storage.clear()` was awaited bare inside a function that promises to isolate
teardown failures and report a count. A rejection there would have escaped —
and the signal handler calls this fire-and-forget, so it would have surfaced
as an unhandled rejection during shutdown, when it is least likely to be seen.
Counted like every other teardown now.

The concurrent-`save()` test did not guarantee both writers loaded before
either saved. If the second `load()` resolved after the first `save()`, it
would read the first turn and even a last-write-wins store would produce both
contents — green without ever reaching the optimistic-conflict merge the test
exists to measure. Both loads are now gated behind a barrier released when the
second arrives.

And a 120-character title with no spaces made the list 835 pixels wide in a
320-pixel viewport, because a flex item's automatic minimum size is its
min-content width. `min-inline-size: 0` plus `overflow-wrap: anywhere`; the
test asserts the page does not scroll sideways rather than asserting the CSS,
so an equivalent fix would keep it green.

lint, check, 82 unit tests, build, and 58 chromium specs pass.
